fix log config
