+++ conftest.py
@@ -1,14 +1,4 @@
 import pytest
-import logging
-
-logging.basicConfig()
-
-pytest_plugins = [
-    'snovault.tests.serverfixtures',
-    'snovault.tests.testappfixtures',
-    'snovault.tests.toolfixtures',
-    'snovault.tests.pyramidfixtures',
-]
 
 def pytest_addoption(parser):
     parser.addoption("--es", action="store", default="", dest='es',
@@ -25,9 +15,3 @@ def remote_es(request):
 @pytest.fixture(scope='session')
 def aws_auth(request):
     return request.config.getoption("--aws-auth")
-
-
-# required so that db transactions are properly rolled back in tests
-@pytest.fixture(autouse=True)
-def autouse_external_tx(external_tx):
-    pass
