+++ src/snovault/app.py
@@ -211,6 +211,7 @@ def main(global_config, **local_config):
     config.include('.authentication')
     config.include('.server_defaults')
     config.include('.types')
+    config.include('.root')
     config.include('.batch_download')
     config.include('.visualization')
